+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Cache/Platform/PlatformCacheTestCreateDestroy.cs
@@ -24,9 +24,9 @@ namespace Apache.Ignite.Core.Tests.Cache.Platform
     using NUnit.Framework;
 
     /// <summary>
-    /// Near cache tests: check create / destroy cache scenarios.
+    /// Platform cache tests: check create / destroy cache scenarios.
     /// </summary>
-    public class CacheNearTestCreateDestroy : IEventListener<CacheEvent>
+    public class PlatformCacheTestCreateDestroy
     {
         /** */
         private IIgnite _grid;
@@ -37,9 +37,6 @@ namespace Apache.Ignite.Core.Tests.Cache.Platform
         /** */
         private IIgnite _client;
         
-        /** */
-        private volatile CacheEvent _lastEvent;
-
         /// <summary>
         /// Fixture set up.
         /// </summary>
@@ -80,88 +77,6 @@ namespace Apache.Ignite.Core.Tests.Cache.Platform
             Ignition.StopAll(true);
         }
 
-        /// <summary>
-        /// Tests the created near cache.
-        /// </summary>
-        [Test]
-        public void TestCreateNearCache(
-            [Values(CacheMode.Partitioned, CacheMode.Replicated)]
-            CacheMode cacheMode,
-            [Values(CacheAtomicityMode.Atomic, CacheAtomicityMode.Transactional)]
-            CacheAtomicityMode atomicityMode)
-        {
-            var cacheName = string.Format("dyn_cache_{0}_{1}", cacheMode, atomicityMode);
-
-            var cfg = new CacheConfiguration(cacheName)
-            {
-                AtomicityMode = atomicityMode,
-                CacheMode = cacheMode
-            };
-
-            var cache = _grid.CreateCache<int, int>(cfg);
-            cache[1] = 1;
-
-            var nearCacheConfiguration = new NearCacheConfiguration();
-            var nearCache = _client.CreateNearCache<int, int>(cacheName, nearCacheConfiguration);
-            Assert.AreEqual(1, nearCache[1]);
-
-            // Create when exists.
-            nearCache = _client.CreateNearCache<int, int>(cacheName, nearCacheConfiguration);
-            Assert.AreEqual(1, nearCache[1]);
-
-            // Update entry.
-            cache[1] = 2;
-            Assert.True(TestUtils.WaitForCondition(() => nearCache[1] == 2, 300));
-
-            // Update through near.
-            nearCache[1] = 3;
-            Assert.AreEqual(3, nearCache[1]);
-
-            // Remove.
-            cache.Remove(1);
-            Assert.True(TestUtils.WaitForCondition(() => !nearCache.ContainsKey(1), 300));
-        }
-
-        /// <summary>
-        /// Tests near cache on the client node.
-        /// </summary>
-        [Test]
-        public void TestCreateNearCacheOnClientNode()
-        {
-            const string cacheName = "client_cache";
-
-            _client.CreateCache<int, int>(cacheName);
-
-            // Near cache can't be started on client node
-            Assert.Throws<CacheException>(
-                () => _client.CreateNearCache<int, int>(cacheName, new NearCacheConfiguration()));
-        }
-
-        /// <summary>
-        /// Tests near cache on the client node.
-        /// </summary>
-        [Test]
-        public void TestCreateCacheWithNearConfigOnClientNode()
-        {
-            const string cacheName = "client_with_near_cache";
-
-            var cache = _client.CreateCache<int, int>(new CacheConfiguration(cacheName),
-                new NearCacheConfiguration());
-
-            AssertCacheIsNear(cache);
-
-            cache[1] = 1;
-            Assert.AreEqual(1, cache[1]);
-
-            var cache2 = _client.GetOrCreateCache<int, int>(new CacheConfiguration(cacheName), 
-                new NearCacheConfiguration());
-
-            Assert.AreEqual(1, cache2[1]);
-
-            cache[1] = 2;
-            Assert.AreEqual(2, cache2[1]);
-        }
-        
         /// <summary>
         /// Tests that near cache data is cleared when underlying cache is destroyed.
         /// </summary>
@@ -304,36 +219,11 @@ namespace Apache.Ignite.Core.Tests.Cache.Platform
         }
         
         /// <summary>
-        /// Asserts the cache is near.
-        /// </summary>
-        private void AssertCacheIsNear(ICache<int, int> cache)
-        {
-            var events = cache.Ignite.GetEvents();
-            events.LocalListen(this, EventType.CacheEntryCreated);
-
-            _lastEvent = null;
-            cache[-1] = int.MinValue;
-
-            TestUtils.WaitForCondition(() => _lastEvent != null, 500);
-            Assert.IsNotNull(_lastEvent);
-            Assert.IsTrue(_lastEvent.IsNear);
-
-            events.StopLocalListen(this, EventType.CacheEntryCreated);
-        }
-
-        /// <summary>
         /// Gets Ignite instance for mode.
         /// </summary>
         private IIgnite GetIgnite(CacheTestMode mode)
         {
             return new[] {_grid, _grid2, _client}[(int) mode];
-        }
-
-        /** <inheritdoc /> */
-        public bool Invoke(CacheEvent evt)
-        {
-            _lastEvent = evt;
-            return true;
         }
         
         /** */
